backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.5.2
pygithub==2.3.0
python-multipart==0.0.9
pydantic-settings==2.2.1
asyncpg==0.29.0
aiosqlite==0.20.0
fastapi-users[sqlalchemy]==13.0.0
fastapi-mail==1.4.1
slowapi==0.1.9
httpx==0.27.0
playwright==1.44.0
pytest-playwright==0.5.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
